Serialize tests that use a common setup/teandown methods in parallel.
pytests parallelizes on the level of a test method + the dimensions. So a test method
which operates on n dimensions will be parallelized n times, and scheduled to run
separately.

A few tests use setup_method/teardown_method in parallel but operate on the same resources
(a test db, test table) etc. This introduces a race when they happen to run at the same
time. This patch forces those tests to run serially, thereby avoiding the race and
removing flakiness.

Change-Id: I77f038e15b9a0616c8df4caaebd31733960ad78e
Reviewed-on: http://gerrit.cloudera.org:8080/207
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: Internal Jenkins