Fix propagation of status through PHJ and PAGG
This was uncovered while investigating IMPALA-1863, but was not the root
cause of that issue.

BufferedTupleStream::AddRow() and AllocateRow() do not return status,
but instead save the status and expect callers to pull the status on the
slow path.  Some callers were ignoring the stream status, which meant
we'd continue forward even when encountering failures.

On the other hand, a second level caller,
PHJ::PartitionProcessProbeBatch() was incorrectly pulling the status
from the stream when PHJ::AppendRow() would have already set
PHJ::status_ (sometimes from the stream, sometimes for other reasons).
So in that case, the failure status can be overwritten by Status::Ok.

This can leave the fragments in an inconsistent state contributing to
the query hangs encountered in IMPALA-1863.

The longer term solution is to fix various shortcoming of the Status
implementation so that the code can simply always return Status
(IMPALA-1948, IMPALA-1916, IMPALA-1986). In the meantime, the above
issues are fixed by passing a Status* for the callee to save the Status
in case of failure.

Change-Id: Ib75b7f88d5ccddfe8448341d5134fa1329c82507
Reviewed-on: http://gerrit.cloudera.org:8080/302
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins