IMPALA-1755: Fix performance regression by improving IN performance
This patch changes how we evaluate IN exprs when all the values in the
IN list are constant (which is quite common). Instead of iterating
through all the values on each input value, we instead create a set of
the constant values in a prepare function, then look up each input
value in this set. In the case where there are non-constant values we
fall back to the old iterative method.

I wrote a benchmark comparing the old and new approach. As expected,
the more values in the IN list, the better the set performs over the
iterative method.

Change-Id: I2ae6b1444a27a87b0a596f96cce3ed3ac7d207ae
Reviewed-on: http://gerrit.cloudera.org:8080/223
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins