IMPALA-1890: Fixing a race between ~BufferedBlockMgr() and the WriteComplete() callback
The DiskIoMgr::HandleWriteFinished() was first decrementing the number of requests in a
disk and then would call the callback (BufferedBlockMgr::WriteComplete()). But the
CancelContext() call would wait until the number of requests per disk was 0. That was
causing a race between the callback and the code after CancelContext(), which was the
destructor of the BufferedBlockMgr. Such a race could cause the problems reported in
IMPALA-1890 and IMPALA-1873.

This patch fixes the problem by reverting the order: the DiskIoMgr first calls the
callback and then decrements the disk request count. It also removes the lock acquisition
after the UnregisterContext() call in the BufferedBlockMgr destructor.

Change-Id: Ib5beea1df68a5ab8d174f197389242d843cd321c
Reviewed-on: http://gerrit.cloudera.org:8080/243
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit eb1b609919b28d12085c1c2aea73a8ccc0699d6c)