IMPALA-1895: FindBuffer() was not rechecking whether it should wait()
The FindBuffer() code in case it had to wait and evict an unpinned buffer it was checking
only once if there are any unpinned buffers and outstanding writes (which means that there
is a reason to wait for a buffer to get freed soon). Then it was entering a loop of
repeatedly calling wait() until there was a free buffer. The problem is that whenever it
was calling wait() it was releasing the lock and another thread could have sneaked in and
claimed the freed buffer. At that point, the thread would call wait() again even though at
this time there was no reason to do so, as there were not unpinned buffers and outstanding
writes. The result most propably would be to "lose" threads for some amount of time, until
all the waiters where woken.

This patch fixes the problem by moving the check whether there is reason to wait() inside
the wait() while loop.

Change-Id: Ie8e9683a32ce39fb3dc3a416e7e3105ef41fc428
Reviewed-on: http://gerrit.cloudera.org:8080/246
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit c38a2a01ea37a4c0795bdb0e0f87abcde4262ff8)