IMPALA-1867: Fix various problems leading to missing MemTracker release()
The general pattern of changes here is that when moving allocations
around, we should not simply "early exit" in case of is_cancelled_ or
errors, but instead put things back into a consistent state so that
cleanup can happen.

Also, in WriteComplete(), I'm reverting the recently added code that
checks for state->is_cancelled() and forces the is_cancelled_ path.
That should not be necessary since we've fixed the race between
WriteComplete() and the destructor, and so it's better to let any
outstanding I/Os finish up normally and the rest of the stack respond
"naturally" to fragment cancellation.

Also, re-enable the consumption() DCHECK that this patch is meant to solve.

Change-Id: I46e7cc046891a656173a743e08d0e31ea720666b
Reviewed-on: http://gerrit.cloudera.org:8080/259
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 1dff448a2c87a7a8077479560e33b3972d202ac9)