HDFS FH Caching: FifoMultimap - aka associative set of stacks (Part II)
The FifoMultimap is a cache-like data structure that keeps `capacity`
number of key-value pairs organized and accessible by key allowing
multiple values per key. If capacity is reached, key-value pairs are
evicted in the least recently added order.

When accessing the values that have identical keys (via Pop(k)), values
are typically returned in LIFO order of this key. However, this property
can only be guaranteed when compiled with C++11 enabled. The motivation
for accessing values in this particular order is to avoid keeping all
values of a particular key hot, even hough only a subset of the values
is actually used.

This class is thread-safe.

Change-Id: I6f1012a5d51fa3d9761796211570fbfa5ae44400
Reviewed-on: http://gerrit.cloudera.org:8080/311
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins