Update/Trim thirdparty/gutil
Working on the integration of Kudu and Impala we're starting to hit
problems due to each project keeping slighlty different versions of
gutil. Simply copy/pasting Kudu's gutil is problematic as it refers
to Kudu often (both in comments and namespace).

Ideally we'd like to pull this to a common place such as the toolchain
Intil we do so, however, the least problematic approach is to trim the
the dependencies to the ones that Impala actually uses and update the
ones that are not trimmed to the latest version, from Kudu, which is
that this patch does.

Change-Id: I935c50344622ff12cde78ef809e7961378ec7774
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6343
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>