Nested types: add ExecNode::Reset() function
The Reset() function resets all data-specific state in the exec node
so that the execution tree can be run again. This will be used by the
subplan operator to run the subplan tree multiple times, once for each
input array.

This patch adds the virtual ExecNode::Reset() function and implements
it for BlockingJoinNode and CrossJoinNode as examples (which are also
required to run a minimal subplan). The other nodes have "NYI" stubs.

Change-Id: I7ec9faa47578d3f97f5e89268ed3b10a124906c5
Reviewed-on: http://gerrit.cloudera.org:8080/329
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins