IMPALA-1955: Remove unnecessary HdfsScanNode::conjunct_ctxs_
During expr refactoring we ended up having a conjunct_ctxs_ at
ExecNode as well as a variable with the same name at HdfsScanNode.
The latter is not needed. We can remove this variable as well as the
work we were doing to create and close the corresponding expr tree
that was never used.

Additionally, it removes an unnecessary variable in hdfs-scan-node
and for disambiguation renames hdfs-scanner's conjunct_ctxs_ to
scanner_conjunct_ctxs_. Also, it makes sure that most variables of
hdfs-scan-node and hdfs-scanner are initiated.

Change-Id: I454491fc3db446066ca155969f84b69cff3b1214
Reviewed-on: http://gerrit.cloudera.org:8080/564
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins