IMPALA-1969: OpenSSL init must not be called concurrently
OpenSSL is not thread-safe. We were previously calling some
initialization code in BufferedBlockMgr::Init() which could
result in concurrent calls to the SSL init code when
multiple queries execute concurrently.

This fixes the issue by moving the initialization to the
(single threaded) impalad initialization so that we don't
have to bother with adding a global lock.

TODO: OpenSSL allows callbacks providing locking to be set
(so that the library is platform agnostic), and in the
future we should set these.

Change-Id: I618a72d6d52dfa812a0617b4f2d6bb384994b87a
Reviewed-on: http://gerrit.cloudera.org:8080/358
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 1cf7103a6699185d407b8ad398b1bc3808889c99)