Add HdrHistogram and HistogramMetric
This patch adds HdrHistogram (ported with minimal changes from an
internal Cloudera port of Gil Tene's original Java HdrHistogram), and
adds a HistogramMetric class to wrap it.

A high-dynamic-range histogram (HdrHistogram) is a histogram
implementation that provides equal precision across the range of
recorded values, bounding the quantization error but keeping memory
requirements low.

By way of illustration, this patch adds two histogram metrics that track
the distribution of query and DDL latencies.

Change-Id: I4aad8ac050cff0feebb6cc2e78d7f97f5155de10
Reviewed-on: http://gerrit.cloudera.org:8080/368
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins