Add the exported kudu client as dependency
This adds the kudu client, exported by running the appropriate build
on the Kudu side, as a dependency in the Impala build. This both makes
sure that a Kudu cmake module exists rooted at KUDU_CLIENT_ROOT and
adds the kudu shared library to the backend.

Note: From this patch on Impala's build will fail if an exported
Kudu client is not present.

Change-Id: I975b7a91be7642a5ad0598dfe521f7077eb304ad
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6619
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>