Add a single-threaded Kudu scan node along with a test
This adds a single-threaded Kudu scan node along with a test.
This first version does not support predicate pushdown and simply
uses a client to scan whatever ranges it got from the frontend.

The test fully verifies that the right data is returned from the
new scan node, in impala's format.

This also adds a method to DescriptorTblBuilder so that we can
set a TTableDescriptor, required in order for it to be usable
in a scan node test.

Finally this moves kudu-util.h to exec/ and adds a .cc file.
The move was required to avoid adding a kudu dependency to util/.

Change-Id: I80b52294d31041232ac728b9f23b17b5cca4ffa5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6654
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins