Add key_columns to TKuduTable and remove col_names
Kudu's client needs a schema in order to perform operations in the server.
KuduScanNodes must thus know the key columns so that they can rebuild a schema,
consistent across all scan nodes, based on the information forwarded from the
frontend.

Note that while it would be apparently possible for the scan nodes to get/cache
the schema from kudu themselves, this would be problematic. The main reason being
that if an alter table operation is undergoing in Kudu, some scan nodes might see
the old schema while other see the new schema, making them return data with
inconsistent schemas.

This patch also removes the 'col_names' field from TKuduTable as that field is
already present in TTable and is thus unnecessary.

Change-Id: I9a36018d57ef95462b28dca6bf3e9e605378debf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6700
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins