Fix key/regular column handling in the frontend's KuduTable
We're currently ignoring the key columns property in the frontend's
KuduTable which stops the backend from correctly rebuilding the Kudu schema.
We're also not setting the column names, which is expected by the backend.

This also changes the nullability defaults so that all key columns
are expected to be non-null and all other columns are nullable. This should
ease compatibility until we have a ddl to specify nullability.

Change-Id: I0c3cde11e52d295083514f47db5cdee83e0d3994
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6818
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>