Handle the case where no slots are materialized
In some cases no slots are materialized (e.g. select count(*) from ....)
but we still need to count the rows.
This was broken for multiple reasons:
- We would build an empty Kudu schema but stil provide a non-zero number
  of keys, causing the Kudu client to crash.
- Next, the scan would fail with a memory exceeded error, caused by us
  interpreting a returned tuple buffer of NULL as an error, versus we actually
  requesting one of size 0 as there is nothing to put there.
- Finally we would still advance row/tuple pointers even when there was
  nothing to advance.

This makes it so that we don't actually allocate any memory or deal
with advancing tuple/row pointers in the case where nothing is materialized.

Finally this changes the scan node test to also test when only some or none
of the slots are materialized.

Change-Id: Ibd27ccd6ae4249d5df52276f79ba49d09b8d3d53
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6830
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>