Analysis of UPDATE
This patch introduces a new feature to SQL: UPDATE. The update syntax is
defined as follows and similar to SQL Server. If a FROM clause is
present the target table must be an alias referencing a table from the
FROM clause. If no FROM clause is present, the table must be a valid
table reference.

  UPDATE [target_alias|target_table_ref]
  SET COL = VAL [, COL = VAL]
  [FROM [table_ref, ...]]
  [WHERE ...]

Currently, updates are only allowed for Kudu tables and disabled for the
other table types. This patch implements the analysis of the update. The
general idea is to follow the same approach as for insert statements.

The parser generates an instance of UpdateStmt that has an embedded
SelectStmt as a query statement. This query statement generates the
values that should be updated and the row keys in the select list. The
where predicate of the update statement is applied similarly to the
select statement.

In addition, this patch renames the Kudu table sink type from KUDU to
KUDU_INSERT, and adds a new sink type called KUDU_UPDATE.

Change-Id: I74b98ecf42cdffb04875dff2dc388344398f5316
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6997
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: jenkins