IMPALA-1702: Dump fragment's thrift descriptor when a table's partition info is corrupt
This patch stores the fragment's TExecPlanFragmentParams, which is the
top-level thrift struct for the fragment, in the runtime state. We
store the fragment params so we can dump everything sent from the FE
to the BE when we detect a problem. This patch does so when a call to
HdfsTable::GetPartition() returns NULL, which we never expect to
happen. This will give us more information in the case of crashes
caused by IMPALA-1702 or similar bugs that cause bad partition IDs to
be sent to the BE.

Change-Id: Ibb2ff05810cfd5f7aa3e210555d9d69361e8272a
Reviewed-on: http://gerrit.cloudera.org:8080/456
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Internal Jenkins