IMPALA-1408: Fix /pprof handlers by allowing web responses with \0 chars
Since we moved to using templates for all our web output, we broke the
/pprof handlers (that can provide profiling information for an Impala
binary as it runs) because the templating library does not deal with
binary data correctly.

This patch allows url callbacks to be registered as 'raw' callbacks
which do not get templatised at all, and just send their output directly
to the HTTP client.

Change-Id: I0c92da3894acfa3e41d35a973961a5c88882b9ed
Reviewed-on: http://gerrit.cloudera.org:8080/463
Tested-by: Internal Jenkins
Reviewed-by: Henry Robinson <henry@cloudera.com>